rename user app
