+++ src/README.md
@@ -23,7 +23,7 @@
                     tests/                                  % holds the tests
                         ...
                     ...
-                myuser/                                     % custom user app
+                user/                                     % custom user app
                     ...
                 templates/
                     registrations/                          % HTML templates for login and logout
