+++ src/application/entrypoint.sh
@@ -3,7 +3,7 @@
 python manage.py check --deploy
 python manage.py makemigrations --noinput
 python manage.py makemigrations main --noinput
-python manage.py makemigrations myuser --noinput
+python manage.py makemigrations user --noinput
 python manage.py migrate --noinput
 python manage.py collectstatic --noinput
 python manage.py add_ssh_fingerprints
