+++ src/application/main/tests/test_models.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from myuser.models import MyUser
+from user.models import MyUser
 
 # Create your tests here.
