+++ src/application/user/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class MyuserConfig(AppConfig):
+class UserConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'myuser'
+    name = 'user'
